added error handler function to avoid littered try/catch code blocks

+++ src/controllers/questions.ts
@@ -1,44 +1,51 @@
 import { Request, Response } from "express";
 import { questionsService } from "../services/questions.js";
-import { QuestionQuery } from "../models/clientModels/question.js";
+import { asyncHandler } from "../utils/asyncHandler.js";
+import { NotFoundError } from "../utils/errors.js";
 
 export const questionsController = {
-  createQuestion: async (req: Request, res: Response) => {
+  createQuestion: asyncHandler(async (req: Request, res: Response) => {
     const payload = req.body;
     const result = await questionsService.createQuestion(payload);
 
     // TODO: Transform the reponse to the required format for the application layer
     // then send that result to the user
     res.status(201).json(result);
-  },
+  }),
 
-  getQuestions: async (req: Request, res: Response) => {
+  getQuestions: asyncHandler(async (req: Request, res: Response) => {
     const queryParams = req.query as any;
     const questions = await questionsService.getAllQuestions(queryParams);
 
     res.status(200).json(questions);
-  },
+  }),
 
-  getQuestion: async (req: Request, res: Response) => {
+  getQuestion: asyncHandler(async (req: Request, res: Response) => {
     const id = req.params.questionId;
     const question = await questionsService.getQuestion(id);
+    if (!question) throw new NotFoundError(`Question with id ${id} not found`);
 
     res.status(200).json(question);
-  },
+  }),
 
-  updateQuestion: async (req: Request, res: Response) => {
+  updateQuestion: asyncHandler(async (req: Request, res: Response) => {
     const id = req.params.questionId;
+    const question = await questionsService.getQuestion(id);
+    if (!question) throw new NotFoundError(`Question with id ${id} not found`);
+
     const payload = req.body;
+    const updatedQuestion = await questionsService.updateQuestion(id, payload);
 
-    const question = await questionsService.updateQuestion(id, payload);
+    res.status(201).json(updatedQuestion);
+  }),
 
-    res.status(201).json(question);
-  },
-
-  deleteQuestion: async (req: Request, res: Response) => {
+  deleteQuestion: asyncHandler(async (req: Request, res: Response) => {
     const id = req.params.questionId;
+    const question = await questionsService.getQuestion(id);
+    if (!question) throw new NotFoundError(`Question with id ${id} not found`);
+
     await questionsService.deleteQuestion(id);
 
     res.status(204).send();
-  },
+  }),
 };
